// dx/frontend/src/api/model/noteOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * dx API
 * OpenAPI spec version: 0.1.0
 */

export interface NoteOut {
  body: string;
  created: string;
  id: string;
  modified: string;
  /** Comma-separated, normalised on write */
  tags: string;
  /** @maxLength 200 */
  title: string;
  version: number;
}